pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "persochattai"
version = "0.1.0"
description = "AI 驅動的英文練習 PWA — Role Play 對話、Podcast 摘要、能力追蹤"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "byoa-core[gemini]",
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "fastrtc>=0.0.28",
    "google-genai>=1.0.0",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.1",
    "beautifulsoup4>=4.14.3",
    "apscheduler>=3.11.0",
    "spacy>=3.8.0",
    "lexical-diversity>=0.1.1",
    "jinja2>=3.1.0",
    "openai>=2.29.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-bdd>=8.1.0",
    "pytest-cov>=7.0.0",
    "pyright>=1.1.408",
    "ruff>=0.14.14",
    "pre-commit>=4.5.1",
]

[tool.uv.sources]
byoa-core = { git = "https://github.com/class83108/Bring-Your-Own-Agent.git" }

# Ruff 配置
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E", "F", "I", "UP",      # pycodestyle, pyflakes, isort, pyupgrade
    "B",                        # bugbear — 抓 Python 常見陷阱
    "SIM",                      # simplify — 強制簡化寫法
    "DTZ",                      # datetimez — timezone 安全
    "T20",                      # print — 禁止 print()
    "N",                        # naming — 命名規範
    "RUF",                      # ruff — 自有 best practice
    "PIE",                      # pie — 移除冗餘
    "C4",                       # comprehensions — Pythonic 寫法
    "RET",                      # return — 乾淨的控制流
]
ignore = [
    "RUF001",                   # 允許中文全形標點出現在字串中
    "RUF002",                   # 允許中文全形標點（專案使用繁體中文）
    "RUF003",                   # 允許中文全形標點出現在註解中
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG", "T20", "E501", "B017"]
"scripts/**/*.py" = ["T20"]

[tool.ruff.format]
quote-style = "single"

# Pyright 配置
[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
include = ["src"]
exclude = ["**/__pycache__", ".venv"]

# Pytest 配置
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = ["-v", "--cov=src/persochattai", "--cov-report=term-missing"]
asyncio_mode = "auto"

# Coverage 配置
[tool.coverage.run]
source = ["src/persochattai"]
branch = true

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:", "if __name__ == .__main__.:"]
